[fix] 動画入出力関係

+++ extensions/BEditor.Extensions.FFmpeg/Encoding/AudioOutputStream.cs
@@ -18,14 +18,16 @@ namespace BEditor.Extensions.FFmpeg.Encoding
         private readonly BinaryWriter _writer;
         private readonly string _videofile;
         private readonly string _pcmfile;
+        private readonly bool _hasVideoStream;
 
-        public AudioOutputStream(AudioEncoderSettings config, string file)
+        public AudioOutputStream(AudioEncoderSettings config, string file, bool hasVideoStream)
         {
             Configuration = config;
             _videofile = file;
             _pcmfile = Path.GetTempFileName();
             _stream = new(_pcmfile, FileMode.Create);
             _writer = new(_stream);
+            _hasVideoStream = hasVideoStream;
         }
 
         public AudioEncoderSettings Configuration { get; }
@@ -47,32 +49,49 @@ namespace BEditor.Extensions.FFmpeg.Encoding
         {
             _stream.Dispose();
             _writer.Dispose();
-
             var ffmpeg = FFmpegExecutable.GetExecutable();
-            var wavfile = Path.ChangeExtension(Path.GetTempFileName(), "wav");
-            var tmpvideo = Path.ChangeExtension(Path.GetTempFileName(), Path.GetExtension(_videofile));
 
-            var process = Process.Start(new ProcessStartInfo(
-                ffmpeg,
-                $"-f f32le -ar {Configuration.SampleRate} -ac 2 -i \"{_pcmfile}\" \"{wavfile}\""))!;
-
-            process.WaitForExit();
-
-            File.Copy(_videofile, tmpvideo);
-            File.Delete(_videofile);
-
-            process = Process.Start(new ProcessStartInfo(
-                ffmpeg,
-                $"-i \"{tmpvideo}\" -i \"{wavfile}\" -c:v copy -c:a aac -map 0:v:0 -map 1:a:0 \"{_videofile}\"")
+            if (_hasVideoStream)
             {
-                CreateNoWindow = true
-            })!;
+                var wavfile = Path.ChangeExtension(Path.GetTempFileName(), "wav");
+                var process = Process.Start(new ProcessStartInfo(
+                    ffmpeg,
+                    $"-f f32le -ar {Configuration.SampleRate} -ac 2 -i \"{_pcmfile}\" \"{wavfile}\"")
+                {
+                    CreateNoWindow = true
+                })!;
 
-            process.WaitForExit();
+                process.WaitForExit();
 
+                var tmpvideo = Path.ChangeExtension(Path.GetTempFileName(), Path.GetExtension(_videofile));
+
+                File.Copy(_videofile, tmpvideo);
+                File.Delete(_videofile);
+
+                process = Process.Start(new ProcessStartInfo(
+                    ffmpeg,
+                    $"-i \"{tmpvideo}\" -i \"{wavfile}\" -c:v copy -c:a aac -map 0:v:0 -map 1:a:0 \"{_videofile}\"")
+                {
+                    CreateNoWindow = true
+                })!;
+
+                process.WaitForExit();
+
+                File.Delete(wavfile);
+                File.Delete(tmpvideo);
+            }
+            else
+            {
+                var process = Process.Start(new ProcessStartInfo(
+                    ffmpeg,
+                    $"-f f32le -ar {Configuration.SampleRate} -ac 2 -i \"{_pcmfile}\" \"{_videofile}\"")
+                {
+                    CreateNoWindow = true
+                })!;
+
+                process.WaitForExit();
+            }
             File.Delete(_pcmfile);
-            File.Delete(wavfile);
-            File.Delete(tmpvideo);
         }
     }
 }
